+++ lib/formats/csv.c
@@ -56,8 +56,8 @@ static void rm_fmt_elem(_UNUSED RmSession *session, _UNUSED RmFmtHandler *parent
        (!file->digest || !session->cfg->write_unfinished)) {
         /* unique file with no partial checksum */
         return;
-        /* TODO: add option to output all unique files */
     }
+
     char checksum_str[rm_digest_get_bytes(file->digest) * 2 + 1];
     memset(checksum_str, '0', sizeof(checksum_str));
     checksum_str[sizeof(checksum_str) - 1] = 0;